requirements.txt
Django==3.2.5
djangorestframework
djangorestframework-simplejwt
django-cors-headers
arrow
django-allauth
dj-rest-auth
drf_yasg
psycopg2
gunicorn
whitenoise[brotli]
python-dotenv
requests
